Added jwt authorisation functions
